requirements.txt
scikit-image
numpy
matplotlib
pandas
pyspark
scipy
seaborn
pillow 
tqdm
tabulate
loguru
